+++ PROYECTO EMILY REP/Assets/Scripts/Save And Load/CharacterSaveData.cs
@@ -26,12 +26,12 @@ namespace KC
         public float zPosition;
 
         [Header("Resources")]
-        public int currentHealth;
+        public float currentHealth;
         public float currentStamina;
 
         [Header("Stats")]
-        public int vitality;
-        public int endurance;
+        public float vitality;
+        public float endurance;
 
         [Header("Sites of Grace")]
         public SerializableDictionary<int, bool> sitesOfGrace;
